# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/client": {
            "get": {
                "description": "deleta um cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client"
                ],
                "summary": "Deleta um cliente",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/client/:id": {
            "get": {
                "description": "Retorna um cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client"
                ],
                "summary": "Lista um cliente",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tenant"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/clients": {
            "get": {
                "description": "Retorna todos os clientes cadastrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client"
                ],
                "summary": "Lista todos os clientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tenant"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/pets": {
            "get": {
                "description": "Retorna todos os pets cadastrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pets"
                ],
                "summary": "Lista todos os pets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/pets/id": {
            "get": {
                "description": "Retorna todos os pets cadastrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pets"
                ],
                "summary": "Lista todos os pets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Users"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Category": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "petshop_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Client": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "description": "Corrigido create_at para created_at",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "observations": {
                    "type": "string"
                },
                "pets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Pet"
                    }
                },
                "petshop_id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Employee": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "string"
                },
                "cellphone": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "created_at": {
                    "description": "UserID    int       `gorm:\"column:user_id;foreignKey:UserID\" json:\"user_id\"`\nUser      User      `gorm:\"foreignKey:UserID\" json:\"usuario\"`",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "petshop_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Modulo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "icone": {
                    "description": "Ícone para o menu",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "ordem": {
                    "description": "Ordem no menu",
                    "type": "integer"
                },
                "rota": {
                    "description": "Rota principal do módulo",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Pet": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "client_id": {
                    "type": "integer"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "observation": {
                    "type": "string"
                },
                "race": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "specie": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "petshop_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Tenant": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/models.WhitelabelConfig"
                },
                "created_at": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TipoAcesso": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "modulos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Modulo"
                    }
                },
                "nome": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Users": {
            "type": "object",
            "properties": {
                "ativo": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "tipo_acesso": {
                    "$ref": "#/definitions/models.TipoAcesso"
                },
                "tipo_acesso_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.WhitelabelConfig": {
            "type": "object",
            "properties": {
                "api": {
                    "type": "object",
                    "properties": {
                        "allowed_origins": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "rate_limit": {
                            "type": "integer"
                        }
                    }
                },
                "branding": {
                    "type": "object",
                    "properties": {
                        "company_name": {
                            "type": "string"
                        },
                        "logo_url": {
                            "type": "string"
                        },
                        "primary_color": {
                            "type": "string"
                        },
                        "secondary_color": {
                            "type": "string"
                        }
                    }
                },
                "features": {
                    "type": "object",
                    "properties": {
                        "enable_feature_x": {
                            "type": "boolean"
                        },
                        "enable_feature_y": {
                            "type": "boolean"
                        },
                        "max_users": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}